requirements.txt
transformers==3.4.0 #4.12.5
# datasets==1.1.3 
scikit-learn==0.23.2
# wandb==0.10.10
pytorch_lightning==1.0.6 #1.6.3